UI for transaction sidebar badge
